hack: use relative paths for api queries
This is done because our server gets exposed under the subpath
`/node0/` through a reverse proxy which forwards the requests to `/`.

This makes it quite difficult to know what the true root of our API
endpoint is at runtime. This is annoying, but since our single-page
web client is exposed at the top level, we can misuse this fact to make
the client's api requests be relative to wherever that is.

Better solution would likely be to configure it with a `config.js` like
file, and template our client side javascript to the correct baseUrl.
But using relative paths also work well here, so its a nice hack.